ユーザページのユーザ情報を「UserInfo.vue」で管理
