+++ plugins/parts.js
@@ -2,10 +2,12 @@ import Vue from 'vue'
 
 import Demo from '../components/Demo'
 import Card from '../components/Card'
+import UserInfo from '../components/UserInfo'
 
 Vue.mixin({
   components: {
     Demo,
     Card,
+    UserInfo,
   }
 })
